chore: add nx to exclude workspaces from build
