--- packages/react/StyledOcticon/index.ts
@@ -1,1 +1,0 @@
-export {default, StyledOcticonProps} from './StyledOcticon'
