--- packages/react/src/StyledOcticon.tsx
@@ -1,18 +1,0 @@
-import {IconProps} from '@primer/octicons-react'
-import React from 'react'
-import styled from 'styled-components'
-import sx, {SxProp} from '../sx'
-import {ComponentProps} from '../utils/types'
-
-type OcticonProps = {icon: React.ElementType; color?: string} & IconProps
-
-function Octicon({icon: IconComponent, ...rest}: OcticonProps) {
-  return <IconComponent {...rest} />
-}
-
-const StyledOcticon = styled(Octicon)<SxProp>`
-  ${({color, sx: sxProp}) => sx({sx: {color, ...sxProp}})}
-`
-
-export type StyledOcticonProps = ComponentProps<typeof StyledOcticon>
-export default StyledOcticon
