+++ packages/react/src/StyledOcticon/index.ts
@@ -1,0 +1,1 @@
+export {default, StyledOcticonProps} from './StyledOcticon'
